votechain/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
requests==2.26.0
cryptography==37.0.4
pycryptodome==3.15.0
flask-cors==3.0.10
python-dotenv==0.19.2
itsdangerous==2.0.1
